requirements.txt
flask>=2.3.0
gunicorn>=21.0.0
gevent>=23.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.24.0
ftfy>=6.1.0
regex>=2023.0.0
git+https://github.com/openai/CLIP.git